rtpstats: RTP_MAX_MISORDER as 3000
RTP_MAX_MISORDER is used to reset stats following RFC3550.
But the point here is that this value (100) are based on
this sentence:
  "Typical values for the parameters are shown, based on a maximum
  misordering time of 2 seconds at 50 packets/second and a maximum
  dropout of 1 minute."

The assumption of this case causes problems when NACKs are used (RFC4585),
because misordering of 100 packets or more can be get.
At this point, stats are reset and some important RR fields as packets_lost
and fraction_lost get not according values.

Signed-off-by: Miguel París Díaz <mparisdiaz@gmail.com>