rtpsource: HACK increment fraction_lost
Without REMB, we have detected that Chrome manages the bitrate (video sent)
using fraction_lost field of RR packets sent by us.
It is also too aggressive increasing the bitrate, so the max BW is overhead
too often and video has a lot of gaps.
If we increase the fraction_lost, Chrome decrements the bitrate more and
the BW overhead is minimized.