gstrtpjitterbuffer: add "rtp-max-dropout" property
This property is useful to config the max lost RTP packets supported.
In automatic mode, the number of packets received in the last seconds
is used to estimate how many losses are supported.

In some cases, when there are a lot of losses is better discard all
packets and notify downstream about it than wait and request
retransmissions because in ,most cases, at the end they don't arrive
(typically when there is bandwidth limitation).